requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
py-bcrypt==0.4
flask-marshmallow==0.14.0
Flask-JWT==0.3.2
pytest==7.1.2